+++ FightCamp-Homework-UIKit/SceneDelegate.swift
@@ -12,17 +12,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let windowScene = (scene as? UIWindowScene) else { return }
 
         window = .init(windowScene: windowScene)
-        window?.rootViewController = HelloViewController() // instantiate the view controller here
+        window?.rootViewController = nil // instantiate the view controller here
         window?.makeKeyAndVisible()
-    }
-}
-
-class HelloViewController: UIViewController {
-
-    override func viewDidLoad() {
-
-        super.viewDidLoad()
-
-        view.backgroundColor = .red
     }
 }
